Allow overriding the View's parameters from the URL
With this, what you define when creating a new chart are the default
parameters. Anything can be overriden by adding it to the query string. For
example, say the view's URL is www.ckan.org/foo/bar. If it's a bar chart, and
you want to turn it into a line chart, just go to
www.ckan.org/foo/bar?chart_type=lines.