Sort the x axis using the order the user defined the filters
We used to only sort the series. Then, say that we got data like:

```
date,region,value
2011,Pakistan,500
2011,Bagh,200
2011,Punjaab,300
```

And my chart is defined as:

* _series_: date
* _xAxis_: region
* _yAxis_: value
* _filters_: "Pakistan", "Bagh"

I would like the order in the x axis to be the same as I filtered. But as
previously we weren't sortering the x axis (just the series), they would end up
alphabetically.

This patch fixes it.

It won't work with horizontal charts, but as we're not supporting them yet,
that's OK.