[#4] Allow multiple filters set through the form
Right now there're 3 hard-coded pairs of filter_field and filter_value, but we
can add more afterwards using JavaScript.

All pairs have the same input name (filter_fields and filter_values). Python
will join them for us and give us two arrays (one for each), with the same
order as defined in the HTML.

With this, we can link the fields to value by their indexes: the 3rd field will
have the 3rd value, etc.

In the JavaScript side, we simply loop through the filter_values and, if
they're different than "", we check which field they belong to and add as a
filter, keeping the order they were defined in.