[#4] Add JS to create/remove filter inputs
I had to create a new JS file for the form because I can't access jQuery when
embedding the JS into <script> tags in the template, because they appear before
jQuery was loaded.

For the view I was able to add it to ```window.onload```, but I can't put
multiple functions there. I'll have to remove that from the view as well,
because we can't rely on being the only JS that uses ```window.onload``` on the
page.

The terrible about it is that, as I need to set a few variables from Python
(the templates, for example), I need to stick them into the global scope.

Desperately needing some refactoring.