Fix bug where multiple charts in the same page became the same kind
The problem was that fields and params were global variables, changed every
time we called init(). So, for every successive call, their contents were
overwritten with the new chart being drawn.

The solution is quite simple, and makes our code better: DON'T USE GLOBALS!

We only using them on the filters, that are constant.