[#14] Set default X-axis ordering when we have no filters
We need to define a default ordering to make sure all series will be in the
same order. For example, imagine that we have a CSV with number of teachers per
region, per year. So we setup a bar chart as:

Series: date
Y Axis: teachers
X Axis: region

If we have more than one date, the data will contain multiple arrays, one for
each. Each of these arrays will have a tuple with [teachers, region]. As we're
ordering these tuples by number of teachers (descending), each date will have a
different ordering. If we set up filters for region, we would fix the order and
we won't have any problems. But if we don't, things will get messy.

With this patch, I sort everything in the same order as the first series'
array.