[#15] Series is optional for bar charts
The attributes' for a bar chart, for example, are:

* X Axis
* Y Axis
* Series

The axes are quite self-explanatory. The series define a kind of "group by".
For example, say that you want to plot the population of every city in the US.
You'll set up the attributes as:

* X Axis == "city"
* Y Axis == "population"
* Series == ???

There's no need for Series in this case. As we required them, you'll end up
setting Series == "city" just to create the chart. This creates another
problem. Now, each bar will be in a different colour, for no reason at all.
Just a lot of chart junk. What you really want is to not set Series.

Series are only useful, in that case, if you'd like to group the bars by state.
So you'd use:

* X Axis == "city"
* Y Axis == "population"
* Series == "state"

Now the bars will be coloured depending on their state, and you'd add a legend
to map which color represent which state.