[#20] Fix bug where charts didn't work with private datasets
The problem is that, by design, the DataStore SQL search doesn't work with
private datasets. We can't use it, then.

I'm using the normal DataStore search, which will only work with a monkey patch
defined in the Pakistan theme (check
https://github.com/okfn/ckanext-pakistan/commit/dcc63f3ffe305f3120c00fe3c73e557953f99b9a#diff-ea271d08933b3bc70fb6e60a02db56bdR244).

That patch changes the DataStore extension to allow multiple filters for the
same attribute. With it, we can filter by an array, as in

  ```{ "region": ["Pakistan", "Punjaab"] }```

Hopefully, we'll merge that in CKAN core when mergin the resource view branch,
so this will work. Both @kindly and I agree that it's a useful change.