Remove mentions to "filter_values" in basiccharts_filters_form
This makes the code more reusable, as you can put in any template you'd like,
and we don't make assumptions on its inputField's name.

We're still assuming that there's one (and only one) input[type="text"][name]
in the template, which is bad, but it's better than it was before.

This change was needed because in ckanext-dashboard we use this code, but with
another "name".