[#12] Workaround for issue with basicGrids' size which would create scrollbars
The problem is that flot ends up drawing over the limits of the canvas
depending on the content. Our workaround is to make the charts a bit smaller
(95% width), so that we have a bit of space to spare, in case it happens.

This doesn't fix the issue, but at least makes it harder to appear.