Respect filters' sort order even if we aren't plotting using its columns
Say that we have the resource:

| id | name | value |
| -- | ---- | ----- |
|  1 | Foo  | 50    |
|  2 | Bar  | 100   |
|  3 | Baz  | 150   |

And we're plotting a bar chart with ```value``` in the Y axis, and ```name```
in the X axis. By default, our chart will sort from the biggest to the smaller.
But if there're filters, it must respect their order.

Right now, if we add filters like:

?filters=name:Foo|name:Bar

It'll create a 2-bars chart with Foo coming first, and Bar second. But if
instead of filtering on ```name```, we filter on ```id```, as:

?filters=id:1|id:2

That didn't overwrote the ordering, so we would still get Bar first, and Foo
second (i.e. ordering by ```value```). This commit changes it.

Now, we order by all the filters, even if we're not using them to draw the
chart.

First, we try to sort by the columns we use on the chart (i.e. group_by,
x_axis, and y_axis), but if there're still ties after that, we use the other
filters.

The order we follow is:

group_by -> x_axis -> y_axis -> everything else, lexicographically sorted

This changes the behavior of vertical bars. Previously, we reversed their
order, so the first elements would appear at the top of the Y axis. Even though
this is visually better, it gives us problems. If we change the order like
that, a dashboard that has horizontal and vertical bar charts using the same
data won't use the same colors for the bars. So Foo might appear yellow in one
chart, and red in other, which is wrong.

The only solution I could find without messing with Flot was keeping the same
sort order.